Add auth-related files and folder structure
